Rewrite crate to properly handle LNURL-Pay/LN-Address
In order to properly resolve LNURL-Pay/LN-Address HRNs, we need to
have the amount the wallet owner wishes to pay ready when we make
the second call to the LNURL-Pay endpoint. Sadly, this requires a
fairly substantial API overhaul.

Instead of being able to simply resolve a string into payment
instructions, we have to resolve our string into an intermediate
state which depends on whether the payment instructions contained
a fixed payment amount.

From there, if no fixed payment amount was set, we require the
developer to make a `set_amount` call which may, among other
things, make the LNURL-Pay callback to fetch a BOLT 11 invoice.

While this is a substantially more complicated API, it still meshes
fairly well with most wallets - after parsing, the UI needs to
decide whether to display an amount selection dialog to the sender
or a simple confirmation screen with a fixed amount. The new API
makes this decision a first-class part of the API.