Use `to_ascii_lowercase` when checking for `lnurl` override start
When calling `instructions.to_lowercase().rfind("lnurl")` and then
using the resulting value as a slice index in `instructions`, we
implicitly assume that any unicode lowercasing will not change the
length of any characters. That's probably fine in practice, but it
seems like a weird assumption to make, so instead we swap it for
`to_ascii_lowercase`.